pyproject.toml
[project]
name = "codex-register-v3-single"
version = "0.1.0"
description = "ChatGPT headless browser auto-registration bot"
requires-python = ">=3.11"
dependencies = [
    "camoufox[geoip]>=0.4.11",
    "playwright>=1.44.0",
    "httpx>=0.27.0",
    "curl-cffi>=0.7.0",
    "aiosqlite>=0.20.0",
    "loguru>=0.7.2",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Setup steps (run manually after uv sync):
#   uv run python -m playwright install firefox chromium
#   uv run python -m camoufox fetch

[tool.uv]
package = true
